s11l103

+++ app/Http/Controllers/JobController.php
@@ -7,15 +7,19 @@ use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\View\View;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class JobController extends Controller
 {
+    use AuthorizesRequests;
+
     /**
      * Display a listing of the resource.
      */
     public function index(): View
     {
-        $jobs = Job::all();
+        $jobs = Job::paginate(9);
 
         return view('jobs.index', compact('jobs'));
     }
@@ -25,6 +29,7 @@ class JobController extends Controller
      */
     public function create(): View
     {
+
         return view('jobs.create');
     }
 
@@ -54,7 +59,7 @@ class JobController extends Controller
             'company_logo' => 'nullable|image|mimes:jpeg,jpg,png,gif|max:2048',
         ]);
 
-        $validatedData['user_id'] = 1;
+        $validatedData['user_id'] = Auth::user()->id;
 
         // Check if a file was uploaded
         if ($request->hasFile('company_logo')) {
@@ -83,6 +88,9 @@ class JobController extends Controller
      */
     public function edit(Job $job): View
     {
+        // Check if the user owns the job
+        $this->authorize('update', $job);
+
         return view('jobs.edit', compact('job'));
     }
 
@@ -91,6 +99,9 @@ class JobController extends Controller
      */
     public function update(Request $request, Job $job): RedirectResponse
     {
+        // Check if the user owns the job
+        $this->authorize('update', $job);
+
         $validatedData = $request->validate([
             'title' => 'required|string|max:255',
             'description' => 'required|string',
@@ -137,12 +148,21 @@ class JobController extends Controller
      */
     public function destroy(Job $job): RedirectResponse
     {
+        // Check if the user owns the job
+        $this->authorize('delete', $job);
+
         // If logo exists, delete it
         if (!empty($job->company_logo)) {
             Storage::disk('public')->delete($job->company_logo);
         }
 
         $job->delete();
+
+        // Check if the request came from the dashboard
+
+        if (request()->query('from') === 'dashboard') {
+            return redirect()->route('dashboard')->with('success', 'Job deleted successfully');
+        }
 
         return redirect()->route('jobs.index')->with('success', 'Job deleted successfully');
     }
